+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable, unreleased changes to this project will be documented in this file.
 
 ### Other changes
 
+- Enhance webhook's subscription query validation. Apply the validation and event inheritance to manifest validation - #11797 by @zedzior
+
 # 3.11.0
 
 ### Highlights